Bump versions for metis and mpi dependencies.
After an upgrade to macOS 13.6 the mpi dependency failed when building ffi bindings. This was solved in a more recent version of bindgen. No release with this bindgen version ins available hence the reference to their repo and to the commit where the last bindgen change in mpi-sys has happened.
As metis has upgraded to a more recent bindgen version in 0.1.2 the fork which we have used before is no longer necessary. And we can point to a regular release.